requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
azure-keyvault
llama-index
llama-index-embeddings-azure-openai
llama-index-llms-azure-openai
pyodbc
urllib3
SQLAlchemy~=2.0.30